+++ packages/d2b-priv-broker/src/runtime.rs
@@ -10520,6 +10520,7 @@ mod tests {
             realm_workloads_launcher_v2_path: None,
             unsafe_local_workloads_path: None,
             provider_registry_v2_path: None,
+            observability_secrets_path: None,
             closures: vec![BundleClosureRef {
                 vm: "corp-vm".to_owned(),
                 path: "closures/corp-vm.json".to_owned(),